requirements.txt
kalshi-python>=1.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
